+++ examples/vector_copy/src/lib.rs
@@ -1,11 +1,8 @@
 #![allow(internal_features)]
-#![feature(abi_gpu_kernel, core_intrinsics, link_llvm_intrinsics)]
+#![feature(abi_gpu_kernel, core_intrinsics, stdarch_amdgpu)]
 #![no_std]
 
-unsafe extern "C" {
-    #[link_name = "llvm.amdgcn.workitem.id.x"]
-    pub safe fn workitem_id_x() -> u32;
-}
+use core::arch::amdgpu;
 
 #[panic_handler]
 fn panic(_: &core::panic::PanicInfo) -> ! {
@@ -15,7 +12,7 @@ fn panic(_: &core::panic::PanicInfo) -> ! {
 #[allow(clippy::missing_safety_doc)]
 #[unsafe(no_mangle)]
 pub unsafe extern "gpu-kernel" fn kernel(input: *const u8, output: *mut u8) {
-    let id = workitem_id_x() as usize;
+    let id = amdgpu::workitem_id_x() as usize;
 
     unsafe {
         // Copy input buffer to output buffer. Each invocation copies one byte.